tests/wpt/tests/element-timing/object.html

<!DOCTYPE HTML>
<head>
<meta charset=utf-8>
<title>Element Timing: observe an object containing an image</title>
<style>
body {
  margin: 20px;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/element-timing-helpers.js"></script>
<script>
  async_test(function (t) {
    let beforeRender;
    let object;
    assert_implements(window.PerformanceElementTiming, "PerformanceElementTiming is not implemented");
    const observer = new PerformanceObserver(
      t.step_func_done(function(entryList) {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        const pathname = window.location.origin + '/images/green-100x50.png';
        checkElement(entry, pathname, 'my_object', 'rectangle', beforeRender, object);
        checkRect(entry, [20, 120, 20, 70]);
        checkNaturalSize(entry, 100, 50);
      })
    );
    observer.observe({entryTypes: ['element']});
    // We add the object during onload to be sure that the observer is
    // registered in time for it to observe the element timing.
    window.onload = () => {
      // Add object of width equal to 100 and height equal to 50.
      object = document.createElement('object');
      object.type = 'image/png';
      object.id = 'rectangle';
      object.data = '/images/green-100x50.png';
      object.setAttribute('elementtiming', 'my_object');

      document.body.appendChild(object);
      beforeRender = performance.now();
    };
  }, 'Element with an object containing an image.');
</script>
</head>
